+++ tests/helpers.py
@@ -108,6 +108,10 @@ def make_test_dataset(
                 np.linspace(0, 1, shape[2], dtype="float64"), dims=dims[2]
             ),
         },
+        attrs={
+            "Conventions": "CF-1.8",
+            "title": f"Test {index + 1}-{index + shape[0]}",
+        },
     )
 
     if crs: